--- pkg/crcontrollers/csibmnode/Dockerfile
@@ -1,7 +1,0 @@
-FROM    ubuntu:20.04
-
-LABEL   description="CSI-Baremetal node CR controller"
-
-ADD     csibmnode  csibmnode-controller
-
-ENTRYPOINT ["/csibmnode-controller"]
