+++ pkg/crcontrollers/operator/Dockerfile
@@ -1,0 +1,7 @@
+FROM    ubuntu:20.04
+
+LABEL   description="Baremetal CSI Operator"
+
+ADD     operator  csi-operator
+
+ENTRYPOINT ["/csi-operator"]
